requirements.txt
flask~=2.2.2
pymysql~=1.1.0
python-dotenv~=0.21.0
pymysqlpool~=1.0.9